+++ portal-service/src/com/liferay/portal/kernel/security/auth/verifier/AuthVerifierConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.auth;
+package com.liferay.portal.kernel.security.auth.verifier;
 
 import java.util.Properties;
